\name{print-methods}
\docType{methods}
\alias{print-methods}

\alias{print,multiclassPairs_object-method}

\alias{print,OnevsrestScheme_genes_SB-method}
\alias{print,OnevsrestScheme_SB-method}

\alias{print,RandomForest_sorted_genes-method}
\alias{print,RandomForest_sorted_rules-method}
\alias{print,rule_based_RandomForest-method}

\title{Methods for Function \code{print} in Package \pkg{multiclassPairs}}
\description{
Methods for function \code{print} in package \pkg{multiclassPairs}
}
\section{Methods}{
\describe{
\item{\code{signature(x = "multiclassPairs_object")}}{}
\item{\code{signature(x = "OnevsrestScheme_genes_SB")}}{}
\item{\code{signature(x = "OnevsrestScheme_SB")}}{}
\item{\code{signature(x = "RandomForest_sorted_genes")}}{}
\item{\code{signature(x = "RandomForest_sorted_rules")}}{}
\item{\code{signature(x = "rule_based_RandomForest")}}{}
}}
\keyword{methods}
